Order and Pizza JPA repository test was added

+++ src/main/java/pizza/repository/card/JpaCardRepository.java
@@ -5,28 +5,24 @@ import java.util.List;
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
 import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.Persistence;
+import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
 
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
 import pizza.domain.AccumulativeCard;
 import pizza.repository.CardRepository;
 
 @Repository(value = "cardRepository")
+@Transactional
 public class JpaCardRepository implements CardRepository {
 	
-	private EntityManagerFactory emf;
-	
+	@PersistenceContext
 	private EntityManager em;
 	
 	public JpaCardRepository() {}
-
-	public JpaCardRepository(EntityManager em) {
-		super();
-		this.em = em;
-	}
 
 	@SuppressWarnings("unchecked")
 	@Override
@@ -38,9 +34,7 @@ public class JpaCardRepository implements CardRepository {
 
 	@Override
 	public AccumulativeCard saveCard(AccumulativeCard card) {
-		em.getTransaction().begin();
 		em.persist(card);
-		em.getTransaction().commit();
 		return card;
 	}
 
@@ -52,31 +46,19 @@ public class JpaCardRepository implements CardRepository {
 	@Override
 	public AccumulativeCard updateAccumulativeCard(AccumulativeCard accumulativeCard) {
 		AccumulativeCard oldCard = em.find(AccumulativeCard.class, accumulativeCard.getId());
-		em.getTransaction().begin();
 		oldCard.setCustomer(accumulativeCard.getCustomer());
 		oldCard.setMoney(accumulativeCard.getMoney());
-		em.getTransaction().commit();
 		return oldCard;
 	}
 
 	@Override
-	public void deleteAccumulativeCard(int id) {
+	public boolean deleteAccumulativeCard(int id) {
 		AccumulativeCard card = em.find(AccumulativeCard.class, id);
-		em.getTransaction().begin();
+		if (card == null) {
+			return false;
+		}
 		em.remove(card);
-		em.getTransaction().commit();
-	}
-	
-	@PostConstruct
-	private void initEntityManager() {
-		emf = Persistence.createEntityManagerFactory("jpa");
-		em = emf.createEntityManager();
-	}
-	
-	@PreDestroy
-	private void closeEntityManager() {
-		em.close();
-		emf.close();
+		return true;
 	}
 
 }
