+++ src/main/java/pizza/repository/customer/JpaCustomerRepository.java
@@ -2,31 +2,25 @@ package pizza.repository.customer;
 
 import java.util.List;
 
-import javax.annotation.PostConstruct;
-import javax.annotation.PreDestroy;
 import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Persistence;
+import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
 
 import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
 
+import pizza.domain.customer.Address;
 import pizza.domain.customer.Customer;
 import pizza.repository.CustomerRepository;
 
+@Transactional
 @Repository(value = "customerRepository")
 public class JpaCustomerRepository implements CustomerRepository {
 	
-	private EntityManagerFactory emf;
-	
+	@PersistenceContext
 	private EntityManager em;
 	
 	public JpaCustomerRepository() {}
-	
-	public JpaCustomerRepository(EntityManager em) {
-		super();
-		this.em = em;
-	}
 
 	@SuppressWarnings("unchecked")
 	@Override
@@ -34,12 +28,10 @@ public class JpaCustomerRepository implements CustomerRepository {
 		Query query = em.createQuery("SELECT customers FROM pizza_service_jpa.customer customers", Customer.class);
 		return query.getResultList();
 	}
-
+	
 	@Override
 	public Customer saveCustomer(Customer customer) {
-		em.getTransaction().begin();
 		em.persist(customer);
-		em.getTransaction().commit();
 		return customer;
 	}
 
@@ -50,32 +42,34 @@ public class JpaCustomerRepository implements CustomerRepository {
 
 	@Override
 	public Customer updateCustomer(Customer customer) {
-		em.getTransaction().begin();
 		Customer oldCustomer = em.find(Customer.class, customer.getId());
-		oldCustomer.setAddresses(customer.getAddresses());
+//		oldCustomer.setAddresses(customer.getAddresses());
 		oldCustomer.setName(customer.getName());
-		em.getTransaction().commit();
 		return oldCustomer;
 	}
 
 	@Override
-	public void deleteCustomer(int id) {
+	public boolean deleteCustomer(int id) {
 		Customer customer = em.find(Customer.class, id);
-		em.getTransaction().begin();
+		if (customer == null) {
+			return false;
+		}
 		em.remove(customer);
-		em.getTransaction().commit();
-	}
-
-	@PostConstruct
-	private void initEntityManager() {
-		emf = Persistence.createEntityManagerFactory("jpa");
-		em = emf.createEntityManager();
+		return true;
 	}
 	
-	@PreDestroy
-	private void closeEntityManager() {
-		em.close();
-		emf.close();
+	@Override
+	public Customer getCustomerWithAddresses(int customerId) {
+		Customer customer = em.find(Customer.class, customerId);
+		customer.getAddresses().size();
+		return customer;
+	}
+
+	@Override
+	public void updateAddresses(Customer customerWithAddresses) {
+		Integer customerId = customerWithAddresses.getId();
+		Customer customer = em.find(Customer.class, customerId);
+		customer.setAddresses(customerWithAddresses.getAddresses());
 	}
 	
 }
